apps/api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
asyncpg==0.30.0
redis==5.2.1
google-genai>=1.0.0
sse-starlette>=2.0.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.1
greenlet==3.1.1
